fix: update Lombok to 1.18.46, fix import paths and type conversion
- backend/pom.xml: bump Lombok 1.18.30→1.18.46, add maven-compiler-plugin annotation processor
- CourseController.java: fix selectCount return type int→Long
- frontend/App.vue: fix import paths to ./src/utils/ and ./src/events/
- frontend/index.html: fix script src /src/main.js→/main.js
- AGENTS.md: add Codex guidance file

+++ backend/src/main/java/com/example/course/controller/CourseController.java
@@ -100,10 +100,10 @@ public class CourseController {
     @RequestMapping("/auto-schedule")
     public Result<String> autoSchedule() {
         if (!"admin".equals(StpUtil.getSession().getString("role"))) return Result.error("无权操作");
-        int pendingCount = courseMapper.selectCount(
+        Long pendingCount = courseMapper.selectCount(
                 new com.baomidou.mybatisplus.core.conditions.query.QueryWrapper<Course>().eq("status", 1));
         String result = autoScheduleService.autoSchedule();
-        eventPublisher.scheduleGenerated(pendingCount > 0 ? pendingCount : 0);
+        eventPublisher.scheduleGenerated(pendingCount > 0 ? pendingCount.intValue() : 0);
         return Result.success(result);
     }
